Add Managers
#Bug:場景轉換間anim會不見

+++ Assets/Scripts/ScriptableObject/LevelDataSO.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 [CreateAssetMenu()]
 public class LevelDataSO : ScriptableObject
 {
-    public int levelTime;
+    public float levelTime;
     public List<int> levelScores;
 }
